% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{AICc}
\alias{AICc}
\title{calculate AICc from a likelihood function optimization procedure}
\usage{
AICc(m)
}
\description{
calculate AICc from a likelihood function optimization procedure
}
