% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.R
\name{akaike_weight_predictions}
\alias{akaike_weight_predictions}
\title{testing : accepts a list of akaike predictions (as returned by akaike_predict
) and uses the corresponding $weight list-item for each model to build
a data.frame of weighted-average predictions.}
\usage{
akaike_weight_predictions(akaike_list = NULL, col = 1)
}
\arguments{
\item{col}{column of the prediction table to average. Typically there are th
ree columns from predict, specifying prediction, upper-bound, and lower-bound}
}
\description{
testing : accepts a list of akaike predictions (as returned by akaike_predict
) and uses the corresponding $weight list-item for each model to build
a data.frame of weighted-average predictions.
}
