% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{akaike_weights}
\alias{akaike_weights}
\title{calculate Akaike weights from a vector of AIC values}
\usage{
akaike_weights(aic_values = NULL, precision = 5)
}
\description{
calculate Akaike weights from a vector of AIC values
}
