% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{balance_zero_transects}
\alias{balance_zero_transects}
\title{testing : drop some zero transects so that zero transects are not greater
than some multiple of our non-zero transects (default is 2X).}
\usage{
balance_zero_transects(imbcr_df = NULL, multiple = 2)
}
\arguments{
\item{multiple}{multiple specifying how many times greater our zero 
transects can be relative to our non-zero transects.}
}
\description{
testing : drop some zero transects so that zero transects are not greater
than some multiple of our non-zero transects (default is 2X).
}
