% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{bartuszevige_intercept_test}
\alias{bartuszevige_intercept_test}
\title{Frequentist slope intercept test first described by Bartuszevige. Does the
confidence interval of a given variable cross the intercept (i.e., x_n=0)?
returns pass/fail by default.}
\usage{
bartuszevige_intercept_test(m = NULL, var = NULL, alpha = 0.975)
}
\arguments{
\item{alpha}{alpha value for our test (default is 0.975)}
}
\description{
Frequentist slope intercept test first described by Bartuszevige. Does the
confidence interval of a given variable cross the intercept (i.e., x_n=0)?
returns pass/fail by default.
}
