% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{binary_reclassify}
\alias{binary_reclassify}
\title{preform a binary reclassification on an input raster, with matching from
values substituted with 1's, and non-matching cells set to NA values. The
output rasters should be thematically consistent with the patch metric
calculations done with the SDMTools package.}
\usage{
binary_reclassify(x = NULL, from = NULL, nomatch = NA)
}
\description{
preform a binary reclassification on an input raster, with matching from
values substituted with 1's, and non-matching cells set to NA values. The
output rasters should be thematically consistent with the patch metric
calculations done with the SDMTools package.
}
