% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{birdcode_fieldname}
\alias{birdcode_fieldname}
\title{hidden function that greps for four-letter-codes}
\usage{
birdcode_fieldname(df = NULL)
}
\description{
hidden function that greps for four-letter-codes
}
