% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{buffer_grid_unit}
\alias{buffer_grid_unit}
\title{generate a square buffer around a single input feature}
\usage{
buffer_grid_unit(x = NULL, row = NULL, radius = 1500)
}
\description{
generate a square buffer around a single input feature
}
