% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{build_unmarked_distance_df}
\alias{build_unmarked_distance_df}
\title{accepts a named raster stack of covariates, an IMBCR SpatialPointsDataFrame,
and a species common name and returns a formatted unmarked distance data.frame
that can be used for model fitting with unmarked.}
\usage{
build_unmarked_distance_df(r = NULL, s = NULL, spp = NULL,
  vars = c("doy", "starttime"), fun = mean, d = c(0, 100, 200, 300,
  400, 500, 600, 700, 800))
}
\description{
accepts a named raster stack of covariates, an IMBCR SpatialPointsDataFrame,
and a species common name and returns a formatted unmarked distance data.frame
that can be used for model fitting with unmarked.
}
