% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{build_unmarked_gds}
\alias{build_unmarked_gds}
\title{accepts a formatted IMBCR SpatialPointsDataFrame and builds an
unmarkedFrameGDS data.frame that we can use for modeling with
the unmarked package.}
\usage{
build_unmarked_gds(df = NULL, numPrimary = 1, distance_breaks = NULL,
  covs = NULL, unitsIn = "m", summary_fun = median,
  drop_na_values = T)
}
\description{
accepts a formatted IMBCR SpatialPointsDataFrame and builds an
unmarkedFrameGDS data.frame that we can use for modeling with
the unmarked package.
}
