% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{buildUnmarkedDistanceDf}
\alias{buildUnmarkedDistanceDf}
\title{accepts a named raster stack, an IMBCR SpatialPointsDataFrame, and a species named
and returns a formatted unmarked distance data.frame that can be used for
model fitting with unmarked.}
\usage{
buildUnmarkedDistanceDf(r = NULL, s = NULL, spp = NULL, vars = c("doy",
  "starttime"), d = c(0, 100, 200, 300, 400, 500, 600, 700, 800))
}
\description{
accepts a named raster stack, an IMBCR SpatialPointsDataFrame, and a species named
and returns a formatted unmarked distance data.frame that can be used for
model fitting with unmarked.
}
