% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_day_of_year}
\alias{calc_day_of_year}
\title{calculate day-of-year for a given IMBCR spatial data.frame}
\usage{
calc_day_of_year(df = NULL)
}
\description{
calculate day-of-year for a given IMBCR spatial data.frame
}
