% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_interpatch_distance}
\alias{calc_interpatch_distance}
\title{hidden function that will calculate interpatch distance using the raster::distance
function}
\usage{
calc_interpatch_distance(x = NULL, stat = mean)
}
\description{
hidden function that will calculate interpatch distance using the raster::distance
function
}
