% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{calc_k}
\alias{calc_k}
\title{estimates a K parameter (upper-bound of integration) as a pre-cursor for
various count-based mixture models (e.g., poisson and negative binomial)}
\usage{
calc_k(df = NULL, multiplier = 1)
}
\description{
estimates a K parameter (upper-bound of integration) as a pre-cursor for
various count-based mixture models (e.g., poisson and negative binomial)
}
