% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_lat_lon}
\alias{calc_lat_lon}
\title{calculate the latitude and longitude of feature centroids in an input dataset
and return an appended Spatial*DataFrame Object with lat/lon entries}
\usage{
calc_lat_lon(s = NULL, proj_str = "+init=epsg:4326")
}
\description{
calculate the latitude and longitude of feature centroids in an input dataset
and return an appended Spatial*DataFrame Object with lat/lon entries
}
