% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_mean}
\alias{calc_mean}
\title{shorthand function for calculating the mean of a raster}
\usage{
calc_mean(x = NULL, na.rm = T)
}
\description{
shorthand function for calculating the mean of a raster
}
