% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_mean_patch_area}
\alias{calc_mean_patch_area}
\title{hidden function that will use SDMTools to calculate the mean patch area
of a given cell}
\usage{
calc_mean_patch_area(x = NULL, area_of_cell = NULL)
}
\description{
hidden function that will use SDMTools to calculate the mean patch area
of a given cell
}
