% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_patch_count}
\alias{calc_patch_count}
\title{hidden function that will use SDMTools to calculate the number of unique patches
on a given raster object}
\usage{
calc_patch_count(x = NULL)
}
\description{
hidden function that will use SDMTools to calculate the number of unique patches
on a given raster object
}
