% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_route_centroids}
\alias{calc_route_centroids}
\title{hidden function that calculates route centroids and attribute the source data
with number of detections for a bird of interest}
\usage{
calc_route_centroids(s = NULL, four_letter_code = NULL,
  use = "detections")
}
\description{
hidden function that calculates route centroids and attribute the source data
with number of detections for a bird of interest
}
