% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_time_of_day}
\alias{calc_time_of_day}
\title{calculate time-of-day for a given IMBCR spatial data.frame}
\usage{
calc_time_of_day(df = NULL)
}
\description{
calculate time-of-day for a given IMBCR spatial data.frame
}
