% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_total_area}
\alias{calc_total_area}
\title{shorthand total area calculation function}
\usage{
calc_total_area(x = NULL, area_of_cell = NULL)
}
\description{
shorthand total area calculation function
}
