% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{calc_transect_effort}
\alias{calc_transect_effort}
\title{calculate transect effort for a given IMBCR spatial data.frame}
\usage{
calc_transect_effort(df = NULL)
}
\description{
calculate transect effort for a given IMBCR spatial data.frame
}
