% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{check_correlation_matrix}
\alias{check_correlation_matrix}
\title{testing :}
\usage{
check_correlation_matrix(var = NULL, x_correlation_matrix = NULL,
  imbcr_df = NULL, correlation_threshold = 0.5)
}
\description{
testing :
}
