% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{check_peasons_product_moment}
\alias{check_peasons_product_moment}
\title{accepts a variable name an an input data.frame and 
conducts a Pearson's product moment correlation test
against all variables in the data.frame.}
\usage{
check_peasons_product_moment(x = NULL, var = NULL)
}
\arguments{
\item{x}{data.frame or matrix containing covariates to test}

\item{var}{variable name (as.character)}
}
\description{
accepts a variable name an an input data.frame and 
conducts a Pearson's product moment correlation test
against all variables in the data.frame.
}
