% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{commonname_fieldname}
\alias{commonname_fieldname}
\title{hidden function that greps for four-letter-codes}
\usage{
commonname_fieldname(df = NULL)
}
\description{
hidden function that greps for four-letter-codes
}
