% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{distance_fieldname}
\alias{distance_fieldname}
\title{hidden function that greps for the distance field name}
\usage{
distance_fieldname(df = NULL)
}
\description{
hidden function that greps for the distance field name
}
