% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{downsample_by_normal_dist}
\alias{downsample_by_normal_dist}
\title{testing : downsample records in a continuous 'distances' distribution using a
normal distribution function (can be truncated). Distances can be literal,
or the output of a Mahalanobis or Euclidean distance function (see: the 'FNN'
package). This function is in testing. It behaves oddly sometimes and 
needs to be refactored. Use with caution.}
\usage{
downsample_by_normal_dist(distances = NULL, bins = 11, shape = 1,
  byid = F, calc_median_bin = F, use_mean = F)
}
\arguments{
\item{shape}{is multiplier applied to the SD of the distances
vector. Multipliers from 1 -> 0 will restrict the variance (increase the
shoulder character of our output).}
}
\description{
testing : downsample records in a continuous 'distances' distribution using a
normal distribution function (can be truncated). Distances can be literal,
or the output of a Mahalanobis or Euclidean distance function (see: the 'FNN'
package). This function is in testing. It behaves oddly sometimes and 
needs to be refactored. Use with caution.
}
