% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{drop_overlapping_units}
\alias{drop_overlapping_units}
\title{hidden function that will calculate the centroids of a USNG
(SpatialPolygons) and return only unique (non-duplicated)
units}
\usage{
drop_overlapping_units(units = NULL)
}
\description{
hidden function that will calculate the centroids of a USNG
(SpatialPolygons) and return only unique (non-duplicated)
units
}
