% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{extract_by_large_df}
\alias{extract_by_large_df}
\title{testing: chunk out the task of extracting buffers from a raster surface
using parallel in the most memory efficient way possible. Useful
for very large (>100,000) list of buffer polygons.}
\usage{
extract_by_large_df(polygon = NULL, r = NULL)
}
\description{
testing: chunk out the task of extracting buffers from a raster surface
using parallel in the most memory efficient way possible. Useful
for very large (>100,000) list of buffer polygons.
}
