% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{extractByTransect}
\alias{extractByTransect}
\title{extract (and optionally, summarize using the fun= argument) raster data across IMBCR transects}
\usage{
extractByTransect(s = NULL, r = NULL, fun = NULL)
}
\description{
extract (and optionally, summarize using the fun= argument) raster data across IMBCR transects
}
