% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{fetchImbcrToLandfireMetadata}
\alias{fetchImbcrToLandfireMetadata}
\title{fetch IMBCR Metadata to Landfire code conversions}
\usage{

  fetchImbcrToLandfireMetadata(url = "https://docs.google.com/spreadsheets/d/1wJ3Xwr67GTYYfim29cKQ9m3AJgoOQryYTHc9v5gJB2g/pub?gid=0&single=true&output=csv")
}
\description{
fetch IMBCR Metadata to Landfire code conversions
}
