% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{gauss_post_stratification}
\alias{gauss_post_stratification}
\title{downsample records in a continuous 'distances' distribution using a
normal distribution function (can be truncated). Distances can be literal,
or the output of a Mahalanobis or Euclidean distance function (see: the 'FNN'
package).}
\usage{
gauss_post_stratification(distances = NULL, bins = 11, shape = 1,
  byid = F, use_median = F)
}
\arguments{
\item{shape}{is multiplier applied to the SD of the distances
vector. Multipliers from 1 -> 0 will restrict the variance (increase the
shoulder character of our output).}
}
\description{
downsample records in a continuous 'distances' distribution using a
normal distribution function (can be truncated). Distances can be literal,
or the output of a Mahalanobis or Euclidean distance function (see: the 'FNN'
package).
}
