% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{gdistsamp_find_optimal_k}
\alias{gdistsamp_find_optimal_k}
\title{testing: there is some ambiguity in selecting a good upper-bounds for integration
in gdistamp (the K parameter). This builds an intercept model around a range
of K values and selects the smallest K value for which a decrease in AIC is no
longer observed. Might also consider checking for stability in the beta parameters
for this as well.}
\usage{
gdistsamp_find_optimal_k(df = NULL, allHabitatCovs = NULL,
  allDetCovs = NULL, mixture = NULL, keyfun = "halfnorm", K = NULL)
}
\description{
testing: there is some ambiguity in selecting a good upper-bounds for integration
in gdistamp (the K parameter). This builds an intercept model around a range
of K values and selects the smallest K value for which a decrease in AIC is no
longer observed. Might also consider checking for stability in the beta parameters
for this as well.
}
