% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{gdistsamp_find_optimal_k_with_transect_downsampling}
\alias{gdistsamp_find_optimal_k_with_transect_downsampling}
\title{testing : given a user-specified mixture distribution, try
and find an optimal K value from a sequence of possible K
values describing the upper-bounds of an integration (using
gdistsamp_find_optimal_k()). Often this will fail if we have
a lot of zeros in our input dataset. This function will attempt
a sequence of different downsampling densities to get around the
zero-inflation problem}
\usage{
gdistsamp_find_optimal_k_with_transect_downsampling(imbcr_df = NULL,
  allHabitatCovs = NULL, allDetCovs = NULL, dist = NULL, K = NULL)
}
\description{
testing : given a user-specified mixture distribution, try
and find an optimal K value from a sequence of possible K
values describing the upper-bounds of an integration (using
gdistsamp_find_optimal_k()). Often this will fail if we have
a lot of zeros in our input dataset. This function will attempt
a sequence of different downsampling densities to get around the
zero-inflation problem
}
