% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{gdistsamp_find_optimal_key_func}
\alias{gdistsamp_find_optimal_key_func}
\title{testing: attempt to fit a series of different detection functions using a user-specified list of detection
covariates. Return the name of the key function that minimizes AIC.}
\usage{
gdistsamp_find_optimal_key_func(imbcr_df = NULL, allDetCovs = NULL)
}
\description{
testing: attempt to fit a series of different detection functions using a user-specified list of detection
covariates. Return the name of the key function that minimizes AIC.
}
