% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{gdistsamp_refit_model}
\alias{gdistsamp_refit_model}
\title{re-fit a model specified by a formula object}
\usage{
gdistsamp_refit_model(formula = NULL, imbcr_df = NULL, K = NULL,
  mixture = NULL, keyfun = NULL)
}
\description{
re-fit a model specified by a formula object
}
