% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{generalizedSingleSeasonAbundance}
\alias{generalizedSingleSeasonAbundance}
\title{Fit the single-season, count-based abundance model of Royle (2004)
to IMBCR data}
\usage{
generalizedSingleSeasonAbundance(t = NULL, spp = NULL)
}
\description{
Fit the single-season, count-based abundance model of Royle (2004)
to IMBCR data
}
