% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{generate_fishnet_grid}
\alias{generate_fishnet_grid}
\title{generate a uniform fishnet grid for a variable polygon dataset containing
one or more geometries. Will fix the rotation of the grid using the bounding
box of the polygon dataset.}
\usage{
generate_fishnet_grid(units = NULL, res = 250)
}
\description{
generate a uniform fishnet grid for a variable polygon dataset containing
one or more geometries. Will fix the rotation of the grid using the bounding
box of the polygon dataset.
}
