% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{imbcr_df_to_spatial_points}
\alias{imbcr_df_to_spatial_points}
\title{parse a source CSV (published by BCR) for IMBCR data and return
as a SpatialPointsDataFrame to the user for post-processing or
conversion to unmarked data.frame objects for modeling}
\usage{
imbcr_df_to_spatial_points(filename = NULL, outfile = NULL,
  write = F, calc_spatial_covs = T)
}
\description{
parse a source CSV (published by BCR) for IMBCR data and return
as a SpatialPointsDataFrame to the user for post-processing or
conversion to unmarked data.frame objects for modeling
}
