% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{imbcr_single_season_pooled_1km_distance}
\alias{imbcr_single_season_pooled_1km_distance}
\title{fit the single-season distance model of Royle (2004) to pooled (station-level)
IMBCR data}
\usage{
imbcr_single_season_pooled_1km_distance(habitatCovs = NULL, detCovs = NULL,
  imbcr_df = NULL)
}
\arguments{
\item{habitatCovs}{vector of habitat covariates to fit}

\item{detCovs}{detection covariates to fit}
}
\description{
fit the single-season distance model of Royle (2004) to pooled (station-level)
IMBCR data
}
