% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{imbcrTableToShapefile}
\alias{imbcrTableToShapefile}
\title{parse a source CSV (published by BCR) for IMBCR data and return as a SpatialPointsDataFrame
to the user}
\usage{
imbcrTableToShapefile(filename = NULL, outfile = NULL, write = F)
}
\description{
parse a source CSV (published by BCR) for IMBCR data and return as a SpatialPointsDataFrame
to the user
}
