% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{l_buffer_grid_units}
\alias{l_buffer_grid_units}
\title{testing: generalization of the buffer_grid_units function that uses
lapply to buffer across all grid features in an input units
data.frame. Parallel implementations for this function are difficult,
because you have to push a copy of units onto each node}
\usage{
l_buffer_grid_units(units = NULL, radius = 1500)
}
\description{
testing: generalization of the buffer_grid_units function that uses
lapply to buffer across all grid features in an input units
data.frame. Parallel implementations for this function are difficult,
because you have to push a copy of units onto each node
}
