% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{l_calc_stat}
\alias{l_calc_stat}
\title{testing: std lapply implementation to benchmark against parLapply}
\usage{
l_calc_stat(x, fun = NULL, from = NULL)
}
\description{
testing: std lapply implementation to benchmark against parLapply
}
