% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{mCombinations}
\alias{mCombinations}
\title{hidden function to derive a data.frame of all possible combinations of vars=}
\usage{
mCombinations(siteCovs = NULL, availCovs = NULL, detCovs = NULL,
  offset = NULL, verbose = T)
}
\description{
hidden function to derive a data.frame of all possible combinations of vars=
}
