% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{monteCarloSpatialHeterogeneity}
\alias{monteCarloSpatialHeterogeneity}
\title{use Monte Carlo randomization to test for heterogeneity in detection across
within-transect spatial replicates}
\usage{
monteCarloSpatialHeterogeneity(s = NULL)
}
\description{
use Monte Carlo randomization to test for heterogeneity in detection across
within-transect spatial replicates
}
