% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{monteCarloTemporalHeterogeneity}
\alias{monteCarloTemporalHeterogeneity}
\title{use Monte Carlo randomization to test for heterogeneity in detection across
within-transect temporal replicates}
\usage{
monteCarloTemporalHeterogeneity(s = NULL, plot = F, main = NA, ...)
}
\description{
use Monte Carlo randomization to test for heterogeneity in detection across
within-transect temporal replicates
}
