% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{par_buffer_grid_units}
\alias{par_buffer_grid_units}
\title{testing: generalization of buffer_grid_unit with parallel comprehension}
\usage{
par_buffer_grid_units(units = NULL, radius = 1500)
}
\description{
testing: generalization of buffer_grid_unit with parallel comprehension
}
