% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.R
\name{par_unmarked_predict}
\alias{par_unmarked_predict}
\title{testing: use the parallel package to predict across a large input
table (with unmarked) using chunking.}
\usage{
par_unmarked_predict(run_table = NULL, m = NULL)
}
\description{
testing: use the parallel package to predict across a large input
table (with unmarked) using chunking.
}
