% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{parseHabitatMetadataByTransect}
\alias{parseHabitatMetadataByTransect}
\title{generate plant community composition data from the associated IMBCR metadata}
\usage{
parseHabitatMetadataByTransect(s, attribute_all = T)
}
\arguments{
\item{attribute_all}{boolean flag indicating whether to take transect-wide metadata and attribute to points}
}
\description{
generate plant community composition data from the associated IMBCR metadata
}
