% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{parseLayerDsn}
\alias{parseLayerDsn}
\title{built-in (hidden) function that will accept the full path of a shapefile 
and parse the string into something that rgdal can understand (DSN + Layer).}
\usage{
parseLayerDsn(x = NULL)
}
\description{
built-in (hidden) function that will accept the full path of a shapefile 
and parse the string into something that rgdal can understand (DSN + Layer).
}
