% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{parseStationCountsAsOccupancy}
\alias{parseStationCountsAsOccupancy}
\title{accepts a list of data.frames (as returned by parseStationLevelMetadata) and post-processes
the counts into binomial detections that can be used to fit occupancy.}
\usage{
parseStationCountsAsOccupancy(detectionHist, na.rm = F)
}
\description{
accepts a list of data.frames (as returned by parseStationLevelMetadata) and post-processes
the counts into binomial detections that can be used to fit occupancy.
}
