% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{parseStationLevelMetadata}
\alias{parseStationLevelMetadata}
\title{parse relavent metadata for estimating p-detection and leave station-level count information in-place
return a list of M data.frames (one for each IMBCR transect) that can be post-processed later for occupancy or abundance
modeling.}
\usage{
parseStationLevelMetadata(s, spp = NULL)
}
\description{
parse relavent metadata for estimating p-detection and leave station-level count information in-place
return a list of M data.frames (one for each IMBCR transect) that can be post-processed later for occupancy or abundance
modeling.
}
