% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{plot_hn_det}
\alias{plot_hn_det}
\title{plot a half-normal detection function from an unmarked object}
\usage{
plot_hn_det(x = NULL, breaks = NULL)
}
\description{
plot a half-normal detection function from an unmarked object
}
