% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{pool_by_transect_year}
\alias{pool_by_transect_year}
\title{hidden function that summarizes imbcr transect covariate data and metadata
by year (with list comprehension). This allows you to calculate covariates
at the IMBCR station level and then pool (summarize) the observations by
transect and year}
\usage{
pool_by_transect_year(x = NULL, df = NULL, breaks = NULL,
  covs = NULL, summary_fun = median)
}
\description{
hidden function that summarizes imbcr transect covariate data and metadata
by year (with list comprehension). This allows you to calculate covariates
at the IMBCR station level and then pool (summarize) the observations by
transect and year
}
