% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{randomWalk_dAIC}
\alias{randomWalk_dAIC}
\title{testing: perform a random walk on an unmarked dataframe with a
user-specified unmarked function and some depth for exploring
our model covariates. This allows for exploring models across
high-dimensional space without having to fit millions of models.
But it's controversial, because Kyle made it up. There are no papers
in the statistics literature demonstrating that this is a valid way
of exploring high-dimensional datasets. So, let's write one.}
\usage{
randomWalk_dAIC(siteCovs = NULL, availCovs = NULL, detCovs = NULL,
  step = 100, umdf = NULL, offset = NULL, depth = 1,
  umFunction = unmarked::distsamp, nCores = NULL, ...)
}
\description{
testing: perform a random walk on an unmarked dataframe with a
user-specified unmarked function and some depth for exploring
our model covariates. This allows for exploring models across
high-dimensional space without having to fit millions of models.
But it's controversial, because Kyle made it up. There are no papers
in the statistics literature demonstrating that this is a valid way
of exploring high-dimensional datasets. So, let's write one.
}
