% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{readOGRfromPath}
\alias{readOGRfromPath}
\title{built-in (hidden) function that will accept the full path of a shapefile and read using rgdal::ogr}
\usage{
readOGRfromPath(path = NULL, verbose = F)
}
\arguments{
\item{path}{argument provides the full path to an ESRI Shapefile}
}
\description{
built-in (hidden) function that will accept the full path of a shapefile and read using rgdal::ogr
}
