% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{rebuildImbcrTable}
\alias{rebuildImbcrTable}
\title{Re-build an IMBCR dataframe for a focal species so that all transect / stations
are represented with their distances or NA values so that the observations
are comprehensible by unmarked. This function is currently in testing.}
\usage{
rebuildImbcrTable(s = NULL, spp = NULL)
}
\description{
Re-build an IMBCR dataframe for a focal species so that all transect / stations
are represented with their distances or NA values so that the observations
are comprehensible by unmarked. This function is currently in testing.
}
