% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{recursive_find_file}
\alias{recursive_find_file}
\title{recursively find all files in folder (root) that match the pattern (name)}
\usage{
recursive_find_file(name = NULL, root = Sys.getenv("HOME"))
}
\description{
recursively find all files in folder (root) that match the pattern (name)
}
