% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{recursiveFindFile}
\alias{recursiveFindFile}
\title{recursively find all files in folder (root) that match the pattern (name)}
\usage{
recursiveFindFile(name = NULL, root = Sys.getenv("HOME"))
}
\description{
recursively find all files in folder (root) that match the pattern (name)
}
