% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{scrub_unmarked_dataframe}
\alias{scrub_unmarked_dataframe}
\title{hidden function used to clean-up an unmarked data.frame (umdf) by dropping
any NA columns attributed by scrub_imbcr_df(), mean-center (scale) our site
covariates (but not sampling effort!), and do some optional quantile filtering
that drops covariates with low variance, which is a useful 'significance
pruning' precursor for principal components analysis. Prefer dropping the NA
bin here (rather than in scrub_imbcr_df), so that we still have an accurate
account of total sampling effort to attribute in scrub_unmarked_dataframe().}
\usage{
scrub_unmarked_dataframe(x = NULL, normalize = T,
  prune_cutoff = NULL)
}
\description{
hidden function used to clean-up an unmarked data.frame (umdf) by dropping
any NA columns attributed by scrub_imbcr_df(), mean-center (scale) our site
covariates (but not sampling effort!), and do some optional quantile filtering
that drops covariates with low variance, which is a useful 'significance
pruning' precursor for principal components analysis. Prefer dropping the NA
bin here (rather than in scrub_imbcr_df), so that we still have an accurate
account of total sampling effort to attribute in scrub_unmarked_dataframe().
}
