% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{simulatedAnnealing_dAIC}
\alias{simulatedAnnealing_dAIC}
\title{use a globally-sensitive numerical optimization procedure to select covariates for
inclusion in our model. This should be considerably faster than randomWalk_dAIC()}
\usage{
simulatedAnnealing_dAIC(m)
}
\description{
use a globally-sensitive numerical optimization procedure to select covariates for
inclusion in our model. This should be considerably faster than randomWalk_dAIC()
}
