% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{singleSeasonAbundance}
\alias{singleSeasonAbundance}
\title{Fit the single-season abundance model of Royle-Nichols (the "RN" model) to IMBCR data}
\usage{
singleSeasonAbundance(x)
}
\description{
Fit the single-season abundance model of Royle-Nichols (the "RN" model) to IMBCR data
}
