% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{singleSeasonDistance}
\alias{singleSeasonDistance}
\title{fit a distance model to IMBCR station data (still testing; doesn't export into namespace yet)}
\usage{
singleSeasonDistance(x = NULL, psi_params = NULL, sigma_params = NULL,
  as_radial_distance = T, link = "half_normal")
}
\description{
fit a distance model to IMBCR station data (still testing; doesn't export into namespace yet)
}
