% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{singleSeasonOccupancy}
\alias{singleSeasonOccupancy}
\title{Fit a single-season occupancy model  assumes a constant probability of species
detection across transects, which is probably inappropriate for IMBCR data and will
lead to inaccurate predictions of occupancy. This is "model m0" from the literature
and loosely follows Andy Royle's (2008) model specification.}
\usage{
singleSeasonOccupancy(parameters, table = NULL, vars = c("a0", "tod", "doy",
  "intensity", "b0", "perc_ag", "perc_grass", "perc_shrub", "perc_tree",
  "perc_playa"))
}
\arguments{
\item{table}{a data.frame of detection histories ('det') and covariates on
occupancy ('b0') and probability of detection ('a0')}

\item{vars}{a vector of strings specifying the covariates used for the analysis.}
}
\description{
Fit a single-season occupancy model  assumes a constant probability of species
detection across transects, which is probably inappropriate for IMBCR data and will
lead to inaccurate predictions of occupancy. This is "model m0" from the literature
and loosely follows Andy Royle's (2008) model specification.
}
