% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{singleSeasonRobustDetectionModel}
\alias{singleSeasonRobustDetectionModel}
\title{Fit the hierarchical abundance model with robust non-detection
estimators for perception vs. detection from Amundson et al. (2014)}
\usage{
singleSeasonRobustDetectionModel(t = NULL, spp = NULL)
}
\description{
Fit the hierarchical abundance model with robust non-detection
estimators for perception vs. detection from Amundson et al. (2014)
}
