% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{singleSeasonTemporaryEmigrationModel}
\alias{singleSeasonTemporaryEmigrationModel}
\title{Fit the hierarchical abundance model of Chandler et al. (2011)}
\usage{
singleSeasonTemporaryEmigrationModel(t = NULL, spp = NULL)
}
\description{
Fit the hierarchical abundance model of Chandler et al. (2011)
}
