% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributing.R
\name{spatial_join}
\alias{spatial_join}
\title{shorthand vector extraction function that performs a spatial join attributes
vector features in x with overlapping features in y. Will automatically
reproject to a consistent CRS.}
\usage{
spatial_join(x = NULL, y = NULL)
}
\description{
shorthand vector extraction function that performs a spatial join attributes
vector features in x with overlapping features in y. Will automatically
reproject to a consistent CRS.
}
