% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{strip_common_name}
\alias{strip_common_name}
\title{strip non-essential characters and spaces from a species common name in a field}
\usage{
strip_common_name(x)
}
\description{
strip non-essential characters and spaces from a species common name in a field
}
