% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{stripCommonName}
\alias{stripCommonName}
\title{strip non-essential characters and spaces from a species common name in a field}
\usage{
stripCommonName(x)
}
\description{
strip non-essential characters and spaces from a species common name in a field
}
