% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{timeperiod_fieldname}
\alias{timeperiod_fieldname}
\title{hidden function that greps for the timeperiod field name}
\usage{
timeperiod_fieldname(df = NULL)
}
\description{
hidden function that greps for the timeperiod field name
}
