% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{transect_fieldname}
\alias{transect_fieldname}
\title{hidden function that greps for the transect field name}
\usage{
transect_fieldname(df = NULL)
}
\description{
hidden function that greps for the transect field name
}
