% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{validateTransectsForSpecies}
\alias{validateTransectsForSpecies}
\title{Validates a raw IMBCR data table for a user-specified species,
NA-filling temporal (minutes) and spatial (station) replicates across
transects where the species was not observed}
\usage{
validateTransectsForSpecies(s = NULL, spp = NULL)
}
\description{
Validates a raw IMBCR data table for a user-specified species,
NA-filling temporal (minutes) and spatial (station) replicates across
transects where the species was not observed
}
