% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{AICc}
\alias{AICc}
\title{calculate AICc from a likelihood function optimization procedure}
\usage{
AICc(m = NULL, p = NULL, n = NULL)
}
\description{
calculate AICc from a likelihood function optimization procedure
}
