% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{mCombinations}
\alias{mCombinations}
\title{testing : hidden function to derive a data.frame of all possible 
combinations of vars=}
\usage{
mCombinations(siteCovs = NULL, availCovs = NULL, detCovs = NULL,
  offset = NULL, verbose = T)
}
\description{
testing : hidden function to derive a data.frame of all possible 
combinations of vars=
}
