% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{mCombinations}
\alias{mCombinations}
\title{derive a data.frame of all possible combinations of vars=}
\usage{
mCombinations(vars = NULL, verbose = T)
}
\description{
derive a data.frame of all possible combinations of vars=
}
