% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kludging.R
\name{parseDetectionsBySpecies}
\alias{parseDetectionsBySpecies}
\title{accepts a SpatialPointsDataFrame of IMBCR data and parses observations into counts for a focal
species specified by the user.}
\usage{
parseDetectionsBySpecies(s, spp = NULL)
}
\arguments{
\item{spp}{character string specifying focal species common name}
}
\description{
accepts a SpatialPointsDataFrame of IMBCR data and parses observations into counts for a focal
species specified by the user.
}
