% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{parseHabitatMetadataByTransect}
\alias{parseHabitatMetadataByTransect}
\title{generate plant community composition data from the associated IMBCR metadata}
\usage{
parseHabitatMetadataByTransect(s)
}
\description{
generate plant community composition data from the associated IMBCR metadata
}
