% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{partialPredict}
\alias{partialPredict}
\title{build a response plot for a given variable using a fitted unmarked model}
\usage{
partialPredict(m = NULL, var = NULL, type = "state", plot = T,
  nCores = NULL, xlab = NULL, ylab = NULL, main = NULL, xlim = NULL,
  ylim = NULL, xTransform = NULL)
}
\description{
build a response plot for a given variable using a fitted unmarked model
}
