% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{permutations}
\alias{permutations}
\title{recursively calculate all possible permutations of an input table n}
\usage{
permutations(n)
}
\description{
recursively calculate all possible permutations of an input table n
}
