% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{randomWalk_dAIC}
\alias{randomWalk_dAIC}
\title{perform a random walk on an unmarked dataframe with a user-specified unmarked funtion}
\usage{
randomWalk_dAIC(vars = NULL, step = 1000, umdf = NULL,
  umFunction = unmarked::distsamp, nCores = NULL, retAll = FALSE)
}
\description{
perform a random walk on an unmarked dataframe with a user-specified unmarked funtion
}
