% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{singleSeasonDistance}
\alias{singleSeasonDistance}
\title{fit a distance model to IMBCR station data (testing)}
\usage{
singleSeasonDistance(psi_params = NULL, sigma_params = NULL,
  in_radial_distance = T)
}
\description{
fit a distance model to IMBCR station data (testing)
}
