% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{singleSeasonOccupancy}
\alias{singleSeasonOccupancy}
\title{Fit a single-season occupancy model  assumes a constant probability of species
detection across transects, which is probably inappropriate for IMBCR data and will
lead to inaccurate predictions of occupancy. This is "model m0" from the literature
and loosely follows Andy Royle's (2008) model specification. It is designed so that
parameter estimates can be derived through an optimization procedure like nlm() and
variables can be easily selected/de-selected, such as for calculating AIC.}
\usage{
singleSeasonOccupancy(det_parameters = NULL, occ_parameters = NULL,
  p = NULL, table = NULL)
}
\arguments{
\item{det_parameters}{proposed parameter values for detection, including 1 intercept (a0)}

\item{occ_parameters}{proposed paramter values for occupancy, including 1 intercept (b0)}

\item{table}{a data.frame of detection histories ('det') and covariates on
occupancy ('b0') and probability of detection ('a0')}
}
\description{
Fit a single-season occupancy model  assumes a constant probability of species
detection across transects, which is probably inappropriate for IMBCR data and will
lead to inaccurate predictions of occupancy. This is "model m0" from the literature
and loosely follows Andy Royle's (2008) model specification. It is designed so that
parameter estimates can be derived through an optimization procedure like nlm() and
variables can be easily selected/de-selected, such as for calculating AIC.
}
