% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbcr_occupancy_modeling.R
\name{validateTransectMetadata}
\alias{validateTransectMetadata}
\title{validate transect-level habitat metadata vs. LANDFIRE}
\usage{
validateTransectMetadata(s)
}
\description{
validate transect-level habitat metadata vs. LANDFIRE
}
